# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gemini/study/flashcards:
    post:
      operationId: generateFlashcards
      tags: [gemini]
      summary: Generate AI flashcards from a topic or text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyInput"
      responses:
        "200":
          description: Generated flashcards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashcardsOutput"
  /gemini/study/quiz:
    post:
      operationId: generateQuiz
      tags: [gemini]
      summary: Generate a quiz from a topic or text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyInput"
      responses:
        "200":
          description: Generated quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizOutput"
  /gemini/study/summarize:
    post:
      operationId: summarizeNotes
      tags: [gemini]
      summary: Summarize notes or text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyInput"
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SummaryOutput"
...
components:
  schemas:
...
    StudyInput:
      type: object
      properties:
        topic:
          type: string
        text:
          type: string
        language:
          type: string
        count:
          type: integer
      required:
        - topic
    Flashcard:
      type: object
      properties:
        front:
          type: string
        back:
          type: string
      required:
        - front
        - back
    FlashcardsOutput:
      type: object
      properties:
        flashcards:
          type: array
          items:
            $ref: "#/components/schemas/Flashcard"
      required:
        - flashcards
    QuizQuestion:
      type: object
      properties:
        question:
          type: string
        options:
          type: array
          items:
            type: string
        correctIndex:
          type: integer
        explanation:
          type: string
      required:
        - question
        - options
        - correctIndex
        - explanation
    QuizOutput:
      type: object
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
      required:
        - questions
    SummaryOutput:
      type: object
      properties:
        summary:
          type: string
        keyPoints:
          type: array
          items:
            type: string
      required:
        - summary
        - keyPoints
...