# lib/api-zod/src/generated/types/quizOutput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuizQuestion } from "./quizQuestion";

export interface QuizOutput {
  questions: QuizQuestion[];
}